% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_obs.R
\name{create_obs}
\alias{create_obs}
\title{Create Observations in Microbiome FHIR report}
\usage{
create_obs(patient_name = "Roel",
  ymd_hms_issued = "2019-08-16T07:03:00Z", snomed_id = NA,
  percent_abundance, species_id)
}
\arguments{
\item{species_id}{}
}
\description{
Create Observations in Microbiome FHIR report
}
