% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_example.R
\name{load_example}
\alias{load_example}
\title{Load example Microbiome data}
\usage{
load_example(name = "LomanNJ_2013.metaphlan_bugs_list.stool",
  all = FALSE)
}
\arguments{
\item{all}{}
}
\description{
Load example Microbiome data
}
