% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json.R
\name{parse_json}
\alias{parse_json}
\title{Parse FHIR report from phyloseq object}
\usage{
parse_json(phyloseq_obj,
  template_path = here("data/fhir_template.RData"))
}
\arguments{
\item{template_path}{}
}
\description{
Parse FHIR report from phyloseq object
}
