% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_template.R
\name{trim_template}
\alias{trim_template}
\title{Trim a FHIR template to add speciment information}
\usage{
trim_template(template)
}
\arguments{
\item{template}{}
}
\description{
Trim a FHIR template to add speciment information
}
